llSPS-INT-3748-Telecom-customer-churn-prediction-using-Watson-Auto-AI
Telecom customer churn prediction using Watson Auto AI

Category: Machine Learning

Skills Required: Python,Python For Data Analysis,Python For Data Visualization,Machine Learning,IBM Cloud,IBM Watson

Project Description:

Customer churn is a major problem and one of the most important concerns for large companies. Due to the direct effect on the revenues of the companies, especially in the telecom field, companies are seeking to develop means to predict potential customer to churn. Therefore, finding factors that increase customer churn is important to take necessary actions to reduce this churn. Churn prediction helps in identifying those customers who are likely to leave a company. The main contribution of our work is to develop a churn prediction model which assists telecom operators to predict customers who are most likely subject to churn. The model developed in this work uses machine learning techniques on IBM platform and builds a new way of features’ engineering and selection.

Solution: Telecommunication industry always suffers from a very high churn rates when one industry offers a better plan than the previous there is a high possibility of the customer churning from the present due to a better plan in such a scenario it is very difficult to avoid losses but through prediction we can keep it to a minimal level. We are building a Machine Learning model to predict the customer churn using IBM Watson AutoAI Machine Learning Service. The model is deployed on IBM cloud to get scoring end point which can be used as API in mobile app or web app building. We are developing a web application which is built using node red service. We make use of the scoring end point to give user input values to the deployed model. The model prediction is then showcased on User Interface.